requirements.txt
Flask==0.10.1
FormEncode==1.2.6
IMDbPY==5.0
Jinja2==2.7.3
MarkupSafe==0.23
SQLAlchemy==0.9.7
SQLObject==1.6.0
Scrapy==0.24.4
Tempita==0.5.2
Twisted==14.0.1
Werkzeug==0.9.6
beautifulsoup4==4.3.2
cffi==0.8.6
cryptography==0.5.4
cssselect==0.9.1
decorator==3.4.0
gunicorn==19.1.1
itsdangerous==0.24
lxml==3.4.0
pbr==0.10.0
pyOpenSSL==0.14
pycparser==2.10
queuelib==1.2.2
six==1.8.0
sqlalchemy-migrate==0.9.2
sqlparse==0.1.11
w3lib==1.10.0
wsgiref==0.1.2
zope.interface==4.1.1
